pr review changes & api docs updated
