--- server/src/constants/feedback.js
@@ -1,12 +1,0 @@
-export const FEEDBACK_STATUS = {
-  PENDING: 'pending',
-  REVIEWED: 'reviewed',
-  RESOLVED: 'resolved',
-  ARCHIVED: 'archived',
-};
-
-export const FEEDBACK_CATEGORY = {
-  ARTICLES: 'articles',
-  CHATS: 'chats',
-  CODE: 'code',
-};
