+++ server/src/controllers/feedback/submit-feedback.js
@@ -8,48 +8,38 @@
  * @returns {Object} Created feedback object
  */
 
-import Feedback from '../../models/feedback.model.js';
+import FEEDBACK from '../../models/feedback.model.js';
 import { sendResponse } from '../../utils/response.js';
 import cloudinary from '../../config/cloudinary.js';
 import { nanoid } from 'nanoid';
-import {
-  FEEDBACK_CATEGORY,
-  FEEDBACK_STATUS,
-} from '../../constants/feedback.js';
+import { FEEDBACK_STATUS, FEEDBACK_CATEGORY } from '../../typings/index.js';
 
-const submitFeedback = async (req, res, next) => {
+const submitFeedback = async (req, res) => {
+  const { title, details, category, reproduce_steps } = req.body;
+  const file = req.file;
+
+  if (!title || title.length < 5 || title.length > 200) {
+    return sendResponse(res, 400, 'Title must be between 5 and 200 characters');
+  }
+  if (!details || details.length < 10 || details.length > 2000) {
+    return sendResponse(
+      res,
+      400,
+      'Details must be between 10 and 2000 characters'
+    );
+  }
+  if (!Object.values(FEEDBACK_CATEGORY).includes(category)) {
+    return sendResponse(res, 400, 'Invalid category');
+  }
+
   try {
-    const { title, details, category, reproduce_steps, attachment } = req.body;
-
-    // Validation
-    if (!title || title.length < 5 || title.length > 200) {
-      return sendResponse(
-        res,
-        400,
-        'Title must be between 5 and 200 characters'
-      );
-    }
-    if (!details || details.length < 10 || details.length > 2000) {
-      return sendResponse(
-        res,
-        400,
-        'Details must be between 10 and 2000 characters'
-      );
-    }
-    if (!Object.values(FEEDBACK_CATEGORY).includes(category)) {
-      return sendResponse(res, 400, 'Invalid category');
-    }
-
     let attachment_url = '';
     let attachment_public_id = '';
 
-    if (attachment) {
+    if (file) {
       try {
-        const date = new Date();
-        const uniqueFileName = `feedback-${nanoid()}-${date.getTime()}`;
-
-        // Upload Base64 image directly to Cloudinary
-        const result = await cloudinary.uploader.upload(attachment, {
+        const uniqueFileName = `feedback-${nanoid()}-${Date.now()}`;
+        const result = await cloudinary.uploader.upload(file.path, {
           public_id: uniqueFileName,
           folder: 'feedback_attachments',
           resource_type: 'image',
@@ -63,9 +53,8 @@ const submitFeedback = async (req, res, next) => {
       }
     }
 
-    const feedback = await Feedback.create({
+    const feedback = await FEEDBACK.create({
       user_id: req.user.user_id,
-
       title,
       details,
       category,
@@ -79,7 +68,8 @@ const submitFeedback = async (req, res, next) => {
       feedback,
     });
   } catch (error) {
-    next(error);
+    console.error('Error submitting feedback:', error);
+    return sendResponse(res, 500, 'Server Error', { error: error.message });
   }
 };
