+++ server/src/models/feedback.model.js
@@ -2,6 +2,6 @@ import { model } from 'mongoose';
 import FEEDBACK_SCHEMA from '../schemas/feedback.schema.js';
 import { COLLECTION_NAMES } from '../constants/db.js';
 
-const Feedback = model(COLLECTION_NAMES.FEEDBACK, FEEDBACK_SCHEMA);
+const FEEDBACK = model(COLLECTION_NAMES.FEEDBACK, FEEDBACK_SCHEMA);
 
-export default Feedback;
+export default FEEDBACK;
