+++ server/src/typings/index.js
@@ -32,3 +32,16 @@ export const COLLABORATION_STATUS = {
   ACCEPTED: 'accepted',
   REJECTED: 'rejected',
 };
+
+export const FEEDBACK_STATUS = {
+  PENDING: 'pending',
+  REVIEWED: 'reviewed',
+  RESOLVED: 'resolved',
+  ARCHIVED: 'archived',
+};
+
+export const FEEDBACK_CATEGORY = {
+  ARTICLES: 'articles',
+  CHATS: 'chats',
+  CODE: 'code',
+};
